requirements.txt
# FarmEye: Smart Farm Monitoring System
# Core web framework
Flask==2.3.3
Werkzeug==2.3.7
click==8.1.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
Flask-Migrate==4.0.5
alembic==1.12.0
pymysql==1.1.0  # For MySQL support in production
psycopg2-binary==2.9.7  # For PostgreSQL support in production

# Authentication and security
Flask-Login==0.6.2
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
Flask-WTF==1.2.1
email-validator==2.0.0.post2
WTForms==3.0.1
PyJWT==2.8.0

# Form handling and validation
WTForms-SQLAlchemy==0.3

# Email
Flask-Mail==0.9.1

# Environment variables
python-dotenv==1.0.0

# File uploads and image processing
Pillow==10.0.1
Flask-Uploads==0.2.1

# API and serialization
Flask-RESTful==0.3.10
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Machine Learning
tensorflow==2.13.0  # For crop/disease detection models
scikit-learn==1.3.1  # For various ML tasks
opencv-python==4.8.1.78  # For image processing
numpy==1.24.3  # Required by TensorFlow and scikit-learn
scipy==1.11.3  # Scientific computing
pandas==2.1.1  # Data manipulation
matplotlib==3.8.0  # Plotting and visualization
joblib==1.3.2  # Model serialization

# GIS and satellite imagery
earthengine-api==0.1.374  # Google Earth Engine API
geojson==3.0.1  # GeoJSON support
pyproj==3.6.1  # Cartographic projections
rasterio==1.3.8  # Raster data processing
geopandas==0.14.0  # Geospatial data operations

# Async tasks and scheduling
celery==5.3.4  # For background tasks like data processing
redis==5.0.1  # For Celery broker
APScheduler==3.10.4  # For scheduled tasks

# Data visualization
plotly==5.17.0  # Interactive charts
dash==2.14.0  # Dashboard components

# AWS integration (for cloud storage)
boto3==1.28.51

# Testing
pytest==7.4.2
pytest-flask==1.2.0
coverage==7.3.2
factory-boy==3.3.0
faker==19.6.2  # For generating test data

# Development tools
Flask-DebugToolbar==0.13.1
black==23.9.1  # Code formatting
flake8==6.1.0  # Code linting
isort==5.12.0  # Import sorting

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0

# Production
gunicorn==21.2.0  # WSGI HTTP Server
supervisor==4.2.5  # Process control